Added route to record a user's expenses
Creating new post route '/:id/addExpense' to allow user to add an expense

Validation of the 'dateofExpense' to ensure it is provided with user input.

Parsing and formatting of the 'dateofExpense' to ISO 8601 format

Retrieval of the bank account using 'accountName' and 'studentId'.

Creation of a new expense record with all provided attributes from the request body