Styled CreatedCard with static details
- Added mo
ck account details to CreateCard component to guide styling display

- Added two SVG assets to impro
ve the UI visuals of the created card for the user.

- Imported CreateCard component to Wallets comp
onent to be reused for making different cards via props.

-Styled div containing card with flex-row
and a gap of 10px to separate the cards and improve the visual display of the list of cards when they ar
e added by the user to the wallets page.